requirements.txt
snowflake-connector-python==3.6.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0

#Why: This is the list of Python packages your project needs.